pyproject.toml
[project]
name = "api-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "WillianCh3"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "motor (>=3.7.1,<4.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pre-commit (==3.7.0)",
    "black (==24.4.2)",
    "isort (==5.12.0)",
    "flake8 (==7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100